Added in body text and reworked some of the sections
I added in body text and removed the section titled "the books" where I had initially intended to provide an in-depth summary of the tv show. However, I realized that this was too much summary for a document that was supposed to be about the concert, rather than the show.